Implement CreateHistoryStream query and refactor the way ports receive RPC requests
This pull request introduces a mechanism by which the application layer
lobs over RPC requests to ports over a channel.

This comes with a refactor of the way RPC requests are handled by the
ports (primary adapters/driving adapters) layer. Usually ports are
driven by external factors and have nothing to do with the application
layer. If that is the case then connections eg. gRPC are fully decoupled
from the application layer and are not driven by the application layer
in any way. This is different in our case as:
- the application layer has to be aware of the concept of connections as
  it must implement a connection manager and initiate connections
- in our case each connections are duplex so the application layer
  actually has to trigger ports as each RPC connection (which the
  application layer can establish) is a source of RPC requests

Due to those reasons it comes as a conclusion that the application layer
has to trigger ports which is allowed as long as a decoupling mechanism
such as pubsub is used. In our case this is simply a channel providing
us with a decoupling mechanism and ensuring concurrent execution.

A bunch of code at the domain level was moved around but wasn't changed.