Cleanup the RPC layer
Response writer is now an interface which makes testing easier.

Removed the stream field from requests as this parameter doesn't seem to
be useful for anything and as far as I can tell it can be guessed when
sending messages.

Moved mux to a separate package to make sure that it is clear that it is
optional and just for convenience and made it use a separate type to
make sure that there is no confusion over which methods to call.

Fixed a bug in listener related to using a nil type when an error was
returned.

Cleaned up response streams but the functionality should not be changed.

Added RPC connection tests.

Incoming request numbers are no longer validated as the protocol
documentation doesn't state anything about this. Added unfortunate
bookkeeping as explained in the comment that I added.

Improved logging by adding a correlation id to the logger for each
connection so that it is easier to follow what is happening in the
context of a single connection. improved raw connection logging to make
sure that the same fields are logged for received and sent messages.

Even though the code in cmd/ is most likely temporary I tried to clean
it up a bit and also made it run the advertiser. I also fixed a bug
which caused the service to not terminate correctly in case of errors
returned by runners.

Apparently something changed with go-ssb's interfaces? I am not sure
what I did but I made sure that go-ssb is at the version that I
confirmed to work correctly.