Fix out-of-order writes in raw connection
If the Write method of RawConnection was called from multiple goroutines
then there was a chance of simultanious writes writing parts of the
message out of order.

I am not adding a test for this as this is a scenario where the test
can't be made completely reliable. I wrote a test which easily
replicated this issue but it is impossible to write a test that is
always guaranteed to detect this problem and doesn't sometimes provide
false negatives. I am therefore not commiting it.