Saves replicated messages in batches
Added a buffer which saves replicated messages in batches.

Previously the main bottleneck when replicating messages was the fact
that each message was saved in a separate transaction. This was made
worse by the fact that Bolt's batch mechanism couldn't really help in
this case as writes were not very concurrent; we were saving messages
sequentially from a couple of workers.

After those changes the main bottleneck should be the slow replication
algorithm which waits doing nothing very often.