Fix long pauses in replication
Because messages do not update the social graph right away (they are
buffered before saving) it is possible that the replication process will
retrieve some messages multiple times. It would be good to fix this in
the future.

There is still a small delay before replication loops as the algorithm
needs to be fixed. Without the pause it would continously ask for all
messages. Instead we should back off if no messages are retrieved for a
feed.