Add very flawed replication manager
Replication manager attempts to replicate a feed from one feed at the
time, prioritizes friends, back offs if no messages are available and
attempts to remember if a peer had more messages for a feed or not.

The manager doesn't cleanup after itself so its memory usage will keep
growing and doesn't have tests.

This is just a quick test to see if this would improve replication
speed.