Add support for recognizing referenced blobs
The code recognizes blobs in two message content types: about and post.

For about messages the code checks the "image" field. This field can be
a string containing a blob ref or a mention object.

For post messages the code checks the mentions field. This field is a
list of mention objects.