Change the want list to work in on demand mode (#15)
After this change the Swift code can request blobs to be downloaded and
the blobs remain in the want list for a specific amount of time. This
replace the previous approach of downloading all blobs right away and
copies the current behaviour exhibited by go-ssb. The new code behaves
slightly better as the want list isn't cleared when application exists
and is persisted.