Add feed tests
I initially didn't write any as I wasn't sure if the interface is there
to stay. For now I don't forsee any changes to it so it is time to add
those tests.

I also made sure that some extra checks were covered and used the
existing function AppendMessage to append created messages for extra
safety.