Make removing feeds from the database possible (#46)
It is now possible to remove an entire feed from the database. When this
is done all the read-models are updated as well, so overall the
following data is removed:
- the mapping between sequence numbers and messages (the feed itself) is removed
- all messages are removed
- data about pubs declared by messages posted to this feed is removed
- entries related to blobs used in messages posted to this feed are removed
- contact entries for this feed are dropped

To achieve this the results of message analysis are passed to the
repositories together with messages. Thanks to that we can track which
data belongs to which message when saving it. This required some domain
changes.