Fix the way contact messages are handled (#58)
Contact messages are now correctly parsed and interpreted. Each contact
message is treated as a set of actions which should be performed on a
contact such as:
  - follow a target
  - unfollow a target
  - block a target
  - unblock a target

One contact message can contain multiple actions eg. block and unfollow.
Follow/unfollow are governed by field "following". Block/unblock are
governed by field "blocking". Fields are mapped as follows:
- following == nil   => []
- following == true  => [follow]
- following == false => [unfollow]
- blocking  == nil   => []
- blocking  == true  => [block]
- blocking  == false => [unblock]

Updates are therefore atomic eg. a message may update one field but
leave the other field untouched.

Some combinations of actions are logically suspicious eg. follow and
block and are treated as errors for now. Message with no actions is
treated as an error.

Based on this each contact effectively has two separate states which are
tracked: whether it is followed and whether it is blocked. This was the
case previously as well but the fields weren't correctly updated.

For example:
    Applying the block operation to contact(following=true, blocking=false)
Results in:
    contact(following=true, blocking=true)

Those fields are not correctly used when building the social graph yet
but they are persisted and updated correctly now.