Cleanup application struct (#113)
The application struct is supposed to define commands and queries which
can be called externally. This is slightly different from normal
approach but this is what we decided at some point. Because of this
commands which are strictly internal were removed. This should reduce
confusion from the bindings perspective.

Added a missing DownloadFeed command to the application struct.