Fix don't send stream close messages for async streams (#115)
Apparently sending a close stream message after an error is received
from a remote for async procedure calls is considered an error. If that
happens peers terminate the connection. So I think the implication is
that you are not allowed to terminate the async streams early either
since that would require sending an error to the remote.