Accelerate replication (#118)
Since the transactions are batched it was previously possible that messages will be dropped from the buffer when the transaction was repeated - the transactions weren't idempotent. This has been fixed.

Additionally previously encountering a malformed known message meant that an error is returned. This has now been changed and malformed messages are persisted without errors. They will have to be rescanned using migrations if the code parsing messages is at fault when the error occurs.

It is still apparently common that the same message is returned from multiple peers which slows down replication. This has to be addressed in the future. Either that or the same peers return the same messages repeatedly. I will continue investigating.