Ignore errors when appending messages to feeds during data migration (#125)
Our task is to import as many messages as possible. We know that there
is invalid data in go-ssb receive logs. Therefore it makes sense to try
to ignore any errors and just try to import as many messages as possible
because otherwise someone is just going to be stuck importing their
data.