Make sure that we are not illegaly reusing badger keys
Badger keys can't be reused without copying them. Matt's database
somehow got corrupted after adding a feed to the ban list. While the
feed data was there all the referenced messages were not. All this data
is removed in transactions so the only explanation that I can think
about is that we are illegally reusing keys which seems to be true.

Added some unrelated tests that I wrote while looking for the bug. They
are unrelated because this bug is hard to test reliably.