Decouple scanning for blobs from known contents (#198)
We need to scan our own messages for blobs in order to push them. The
previous approach to finding blobs in messages was not going to scale.
This commit prepares for this by improving the way messages are scanned
for blobs.

Previously we needed to manually implement every known message content
type that supported blobs. Now we simply parse the entire content
looking for blob refs.

This also decouples scanning for blob refs from known message content
types which is great.