Fix receive log data corruption on migration (#206)
Ignored messages were treated as successes and put in the receive log. The receive log would then break because those messages didn't actually exist in the database.

Fixes https://github.com/planetary-social/scuttlego/issues/200.

In reality we should solve this at the domain level and this bug stems from the fact that we didn't.