Detect and handle forks during migration
Sometimes messages are quietly forked and we don't know which is the
correct one. If we detect a message conflicts with an existing message
we drop all messages with sequence >= sequence of that message so let
the conflic resync from the network.

Related to https://github.com/planetary-social/scuttlego/issues/197.